Improved version

+++ numba/core/caching.py
@@ -106,11 +106,16 @@ class _CacheLocator(metaclass=ABCMeta):
         path = self.get_cache_path()
         os.makedirs(path, exist_ok=True)
         # Ensure the directory is writable by trying to write a temporary file
-        # We avoid using tempfile.TemporaryFile since it may hang indefinitely on Windows:
-        # https://github.com/python/cpython/issues/66305
-        filename = os.path.join(path, "cache_write_test.txt")
-        open(filename, mode="wb").close()
-        os.remove(filename)
+        if os.name == "nt":
+            # We avoid using tempfile.TemporaryFile or tempfile.mkstemp since it may hang indefinitely on Windows:
+            # https://github.com/python/cpython/issues/66305
+            tmpfile = tempfile.mktemp(dir=path)
+            open(tmpfile, mode="wb").close()
+            os.remove(filename)
+        else:
+            # on other operating systems it's safer to rely on
+            # TemporaryFile.
+            tempfile.TemporaryFile(dir=path).close()
 
     @abstractmethod
     def get_cache_path(self):
